DPTS-558 Refactor execute_db_query to database
